Software Spinepose/requirements.txt
ultralytics>=8.0.0
mediapipe==0.10.14
opencv-python>=4.8.0
numpy>=1.24.0
google-genai>=0.3.0
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart
